backend/requirements.txt
# API
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9

# DB — SQLite with async support
sqlalchemy==2.0.30
aiosqlite==0.20.0
asyncpg==0.29.0

# Migrations
alembic==1.13.2

# Auth
python-jose[cryptography]==3.3.0
bcrypt==4.1.3

# Parsing
pdfplumber==0.11.0
python-docx==1.1.2

# ML / Embeddings
sentence-transformers==3.0.1
numpy==1.26.4
scikit-learn==1.5.0

# LLM
langchain==0.2.5
langchain-openai==0.1.8
openai==1.30.5
langsmith>=0.1.0

# Validation
pydantic==2.7.3
pydantic-settings==2.3.1

# Utils
python-dotenv==1.0.1
httpx==0.27.0
tenacity==8.3.0
structlog==24.2.0

# Dev / Testing
pytest==8.2.2
pytest-asyncio==0.23.7


# Gemini LLM
google-generativeai>=0.5.4

# NLP
spacy>=3.7.0

# Additional utils
python-dateutil==2.9.0